Backend/requirements.txt
flask
requests
python-dotenv
firebase-rest-api